requirements.txt
discord-py-interactions
git+https://github.com/lukect/undetected-chromedriver.git
pickle
pyyaml
